Step 43

+++ src/output.zig
@@ -26,7 +26,7 @@ pub fn editorDrawRows(ab: *abuf.abuf) !void {
             const welcome:[] u8 = try std.fmt.allocPrint(std.heap.page_allocator,
                 "{s} Editor -- version: {s}",
                 .{data.editorName, data.version});
-            var padding:u16 = (data.editor.screenCols - welcome.len) / 2;
+            var padding:usize = (data.editor.screenCols - welcome.len) / 2;
             if(padding != 0){
                 try ab.append("~");
                 padding -= 1;
